blog_backend/requirements.txt
django==3.0.1
djangorestframework==3.11.1
itsdangerous==1.1.0
django-cors-headers==3.5.0
mysqlclient==2.0.1
django-redis==4.12.1
psycopg2==2.8.6
Pillow==4.1.0
django-filter==2.4.0
djangorestframework-jwt==1.11.0
requests==2.24.0
drf-yasg==1.20.0
tencentcloud-sdk-python==3.0.316
